Go5

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminRestController.java
@@ -1,6 +1,8 @@
 package ru.kata.spring.boot_security.demo.controller;
 
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -28,46 +30,54 @@ public class AdminRestController {
         this.userService = userService;
     }
 
-    @GetMapping("/")
-    public ModelAndView printWelcome() {
-        ModelAndView mav = new ModelAndView("/index");
-        mav.addObject("index", "index");
-        return mav;
-    }
-
     @GetMapping("/admin-js")
     public ModelAndView showAllUsers() {
-        ModelAndView mav = new ModelAndView("/admin-js");
+        ModelAndView mav = new ModelAndView("admin-js");
         mav.addObject("adminUrl", adminUrl);
         mav.addObject("userUrl", userUrl);
         mav.addObject("admin-js", "admin-js");
         return mav;
     }
 
+
     @GetMapping("/api/admin")
-    public List<User> findAllUsers() {
-        return userService.getAllUsers();
+    public ResponseEntity<List<User>> findAllUsers() {
+        List<User> users = userService.getAllUsers();
+        return users.isEmpty()
+                ? ResponseEntity.noContent().build()
+                : ResponseEntity.ok(users);
     }
 
     @GetMapping("/api/admin/{id}")
-    public User findUser(@PathVariable("id") int id) {
-        return userService.getUserById(id);
+    public ResponseEntity<User> findUser(@PathVariable int id) {
+        User user = userService.getUserById(id);
+        return user != null
+                ? ResponseEntity.ok(user)
+                : ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/api/admin/{id}")
-    public void deleteUser(@PathVariable("id") int id) {
-        userService.removeUserById(id);
+    public ResponseEntity<Void> deleteUser(@PathVariable int id) {
+        if (userService.getUserById(id) != null) {
+            userService.removeUserById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @PostMapping("/api/admin")
-    public User createUser(@RequestBody User user) {
+    public ResponseEntity<User> createUser(@RequestBody User user) {
         userService.addUser(user);
-        return user;
+        return ResponseEntity.status(HttpStatus.CREATED).body(user);
     }
 
     @PatchMapping("/api/admin/{id}")
-    public User updateUser(@RequestBody User updateUser, @PathVariable("id") int id) {
+    public ResponseEntity<User> updateUser(@RequestBody User updateUser, @PathVariable int id) {
+        User existingUser = userService.getUserById(id);
+        if (existingUser == null) {
+            return ResponseEntity.notFound().build();
+        }
         userService.updateUser(id, updateUser);
-        return findUser(id);
+        return ResponseEntity.ok(userService.getUserById(id));
     }
 }
